# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Book {
  id: number;
  title: string;
  /** @nullable */
  author: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverColor?: string | null;
  chapterCount: number;
  pageCount: number;
  createdAt: string;
  updatedAt: string;
}

export type ChapterType = typeof ChapterType[keyof typeof ChapterType];


export const ChapterType = {
  chapter: 'chapter',
  subchapter: 'subchapter',
  section: 'section',
} as const;

export interface Chapter {
  id: number;
  bookId: number;
  title: string;
  type: ChapterType;
  content: string;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface BookDetail {
  id: number;
  title: string;
  /** @nullable */
  author: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverColor?: string | null;
  createdAt: string;
  updatedAt: string;
  chapters: Chapter[];
}

export interface BookInput {
  /** @minLength 1 */
  title: string;
  author?: string;
  description?: string;
  coverColor?: string;
}

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  author?: string;
  description?: string;
  coverColor?: string;
}

export type ChapterInputType = typeof ChapterInputType[keyof typeof ChapterInputType];


export const ChapterInputType = {
  chapter: 'chapter',
  subchapter: 'subchapter',
  section: 'section',
} as const;

export interface ChapterInput {
  /** @minLength 1 */
  title: string;
  type?: ChapterInputType;
  content: string;
  sortOrder?: number;
}

export type ChapterUpdateType = typeof ChapterUpdateType[keyof typeof ChapterUpdateType];


export const ChapterUpdateType = {
  chapter: 'chapter',
  subchapter: 'subchapter',
  section: 'section',
} as const;

export interface ChapterUpdate {
  /** @minLength 1 */
  title?: string;
  type?: ChapterUpdateType;
  content?: string;
  sortOrder?: number;
}

export interface ChapterReorder {
  orderedIds: number[];
}

export interface AppendTextInput {
  text: string;
  chapterTitle?: string;
}

export interface ParseTextInput {
  text: string;
}

export type ParsedChapterType = typeof ParsedChapterType[keyof typeof ParsedChapterType];


export const ParsedChapterType = {
  chapter: 'chapter',
  subchapter: 'subchapter',
  section: 'section',
} as const;

export interface ParsedChapter {
  title: string;
  type: ParsedChapterType;
  content: string;
}

export interface ParseResult {
  suggestedTitle: string;
  chapters: ParsedChapter[];
}

export interface LibraryStats {
  totalBooks: number;
  totalChapters: number;
  totalPages: number;
  recentBooks: Book[];
}