# lib/api-zod/src/generated/types/appendTextInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */

export interface AppendTextInput {
  text: string;
  chapterTitle?: string;
}