# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */

export interface Book {
  id: number;
  title: string;
  /** @nullable */
  author: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverColor?: string | null;
  chapterCount: number;
  pageCount: number;
  createdAt: Date;
  updatedAt: Date;
}