# lib/api-zod/src/generated/types/bookDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { Chapter } from './chapter';

export interface BookDetail {
  id: number;
  title: string;
  /** @nullable */
  author: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverColor?: string | null;
  createdAt: Date;
  updatedAt: Date;
  chapters: Chapter[];
}