# lib/api-zod/src/generated/types/bookUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  author?: string;
  description?: string;
  coverColor?: string;
}