# lib/api-zod/src/generated/types/chapter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterType } from './chapterType';

export interface Chapter {
  id: number;
  bookId: number;
  title: string;
  type: ChapterType;
  content: string;
  sortOrder: number;
  createdAt: Date;
  updatedAt: Date;
}