# lib/api-zod/src/generated/types/chapterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterInputType } from './chapterInputType';

export interface ChapterInput {
  /** @minLength 1 */
  title: string;
  type?: ChapterInputType;
  content: string;
  sortOrder?: number;
}