# lib/api-zod/src/generated/types/chapterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ChapterUpdateType } from './chapterUpdateType';

export interface ChapterUpdate {
  /** @minLength 1 */
  title?: string;
  type?: ChapterUpdateType;
  content?: string;
  sortOrder?: number;
}