# lib/api-zod/src/generated/types/libraryStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from './book';

export interface LibraryStats {
  totalBooks: number;
  totalChapters: number;
  totalPages: number;
  recentBooks: Book[];
}