# lib/api-zod/src/generated/types/parseResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedChapter } from './parsedChapter';

export interface ParseResult {
  suggestedTitle: string;
  chapters: ParsedChapter[];
}