# lib/api-zod/src/generated/types/parsedChapter.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Text-to-Book Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedChapterType } from './parsedChapterType';

export interface ParsedChapter {
  title: string;
  type: ParsedChapterType;
  content: string;
}